+++ resources/views/back/pages/settings.blade.php
@@ -39,12 +39,25 @@
                         <div class="mb-2" style="max-width: 200px">
                             <img src="" alt="" class="img-thumbnail" id="logo-img-preview" data-ijabo-default-img="{{ \App\Models\Setting::find(1)->blog_logo }}" >
                         </div>
-                        <form action="{{ route('author.change-blog-logo')}}" method="POST" id="changeBlogLogoForm" >
+                        <form action="{{ route('author.change-blog-logo') }}" method="POST" id="changeBlogLogoForm" >
                             @csrf
                             <div class="mb-2" >
                                 <input type="file" name="blog_logo" class="form-control">
                             </div>
                             <button class="btn btn-primary">Change logo</button>
+                        </form>
+                    </div>
+                    <div class="col-md-6" >
+                        <h3>Set blog favicon</h3>
+                        <div clas="mb-2" style="max-width: 100px" >
+                            <img src="" class="img-thumbnail" id="favicon-image-preview"  data-ijabo-default-img="{{ \App\Models\Setting::find(1)->blog_favicon }}" >
+                        </div>
+                        <form action="{{ route('author.change-blog-favicon') }}" method="POST" id="changeBlogFaviconForm" >
+                            @csrf
+                            <div class="mb-2" >
+                                <input type="file" name="blog_favicon" class="form-control">
+                            </div>
+                            <button class="btn btn-primary">Change favicon</button>
                         </form>
                     </div>
                 </div>
@@ -76,7 +89,22 @@
         }
     });
 
-    $('#changeBlogLogoForm').submit( function(e) {
+    $('input[name="blog_favicon"]').ijaboViewer( {
+        preview         : '#favicon-image-preview',
+        imageShape      : 'square',
+        allowedExtensions : ['ico'],
+        onErrorShape : function(message, element) {
+            alert(message+'=> width should be equal to height');  // \u{2260} is html symbol ≠	&#8800;	&#x2260; &ne; NOT EQUAL TO
+        },
+        onInvalidType : function(message, element) {
+            alert(message);
+        },
+        onSuccess : function(message, element) {
+            // alert(message);
+        }
+    } );
+
+    $("#changeBlogLogoForm").submit( function(e) {
         e.preventDefault();
         var form = this;
         $.ajax({
@@ -100,6 +128,30 @@
         });
     } );
     
+    $("#changeBlogFaviconForm").submit( function(e) {
+        e.preventDefault();
+        var form = this;
+        $.ajax({
+            url         : $(form).attr('action'),
+            method      : $(form).attr('method'),
+            data        : new FormData(form),
+            processData : false,
+            dataType    : 'json',
+            contentType : false,
+            beforeSend  : function() {},
+            success     : function(data){
+                toastr.remove();
+                if (data.status == 1) {
+                    toastr.success(data.msg);
+                    $(form)[0].reset();
+                    Livewire.dispatch('updateTopHeader');
+                } else {
+                    toastr.error(data.msg);
+                }
+            }
+        });
+    } );
+
   </script>
 
 @endpush